organize screen files
